requirements.txt
aniso8601==9.0.1
argcomplete==1.12.3
attrs==21.4.0
bagit==1.8.1
CacheControl==0.12.6
certifi==2021.5.30
charset-normalizer==2.0.5
click==8.0.1
coloredlogs==15.0.1
cwl-upgrader==1.2
cwl-utils==0.11
cwlformat==2021.1.5
cwltool==3.1.20210816212154
Flask==1.1.1
Flask-Cors==3.0.8
Flask-JWT==0.3.2
flask-restplus==0.13.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
Flask-Testing==0.7.1
greenlet==1.1.2
humanfriendly==9.2
idna==3.2
isodate==0.6.0
itsdangerous==2.0.1
Jinja2==3.0.1
jsonschema==4.4.0
lockfile==0.12.2
lxml==4.6.3
MarkupSafe==2.0.1
mistune==0.8.4
msgpack==1.0.2
mypy-extensions==0.4.3
mysqlclient==1.4.6
networkx==2.6.3
prov==1.5.1
psutil==5.8.0
pydot==1.4.2
PyJWT==1.4.2
pyparsing==2.4.7
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2021.3
rdflib==6.0.1
rdflib-jsonld==0.6.1
requests==2.26.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.6
schema-salad==8.2.20210914115719
shellescape==3.8.1
six==1.16.0
SQLAlchemy==1.4.29
typing==3.7.4.3
typing-extensions==3.10.0.2
urllib3==1.26.6
# uWSGI==2.0.19.1
Werkzeug==0.16.0
